Customer, Courier APIs created | Business Services Created | KeyCloak integrated | Exception Handling, Logging Configured

+++ ms-user/src/main/java/org/company/user/domain/entities/Courier.java
@@ -3,12 +3,9 @@ package org.company.user.domain.entities;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.company.user.domain.CustomerLevel;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
@@ -26,7 +23,7 @@ public class Courier implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "FIRST_NAME", nullable = false)
     private String firstName;
